refactor: #23 Move HttpClientUtils to non-sparql package
