+++ cogni-libs/src/main/java/zone/cogni/libs/sparqlservice/impl/StardogSparqlService.java
@@ -1,7 +1,7 @@
 package zone.cogni.libs.sparqlservice.impl;
 
 import static com.google.common.net.HttpHeaders.CONTENT_TYPE;
-import static zone.cogni.libs.sparqlservice.impl.HttpClientUtils.execute;
+import static zone.cogni.libs.core.utils.HttpClientUtils.execute;
 
 import java.io.File;
 import java.io.FileNotFoundException;
@@ -18,6 +18,7 @@ import org.apache.jena.rdf.model.Model;
 import org.apache.jena.riot.Lang;
 import org.apache.jena.sparql.exec.http.QueryExecutionHTTPBuilder;
 import org.springframework.http.MediaType;
+import zone.cogni.libs.core.utils.HttpClientUtils;
 import zone.cogni.libs.jena.utils.JenaUtils;
 import zone.cogni.libs.jena.utils.TripleSerializationFormat;
 import zone.cogni.libs.sparqlservice.SparqlService;
